Move Mention and MessageMember to common.json
